Refactor in subbundles
This way the TwigList part can easily be deactivated.
As that part is only usefull pre integration!

FrontendBundle can contain anything necessary to communicate
with elasticMS datastructures in ElasticSearch
