+++ .gitignore
@@ -1,3 +1,4 @@
+.idea
 /.sass-cache/
 /.settings/
 .DS_Store
