--- EMSClientHelperBundle.php
@@ -1,9 +1,0 @@
-<?php
-
-namespace EMS\ClientHelperBundle;
-
-use Symfony\Component\HttpKernel\Bundle\Bundle;
-
-class EMSClientHelperBundle extends Bundle
-{
-}
