--- Resources/config/config.yml
@@ -1,1 +1,0 @@
-ems_client_helper:    twig_list :        templates:              - {resource: 'EMSClientHelperBundle', path: 'Resources/views/Pages'}        app_enabled: false        app_base_path : #['Resources/views/default']
