--- Resources/views/Elements/footer.html.twig
@@ -1,11 +1,0 @@
-{# This file is part of the EMSClientHelperBundle #}
-
-<footer class="footer">
-  <div class="container">
-  	   Footer 
-  	   	<div class="info">
-		<p>'This is the footer template. Feel free to override it.' </p>
-		<p>To do so, find the original file in ./views/Elements/footer.html.twig <br/> and copy it to your project's <span>app/Resources/EMSClientHelperBundle/views/</span> folder whith respect to the original folder structure</p>
-   	</div>
-  </div>
-</footer>
