--- Tests/Controller/DefaultControllerTest.php
@@ -1,17 +1,0 @@
-<?php
-
-namespace EMS\ClientHelperBundle\Tests\Controller;
-
-use Symfony\Bundle\FrameworkBundle\Test\WebTestCase;
-
-class DefaultControllerTest extends WebTestCase
-{
-    public function testIndex()
-    {
-        $client = static::createClient();
-
-        $crawler = $client->request('GET', '/');
-
-        $this->assertContains('Hello World', $client->getResponse()->getContent());
-    }
-}
