+++ WebClient/FrontendBundle/FrontendBundle.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace EMS\ClientHelperBundle\WebClient\Frontend;
+
+use Symfony\Component\HttpKernel\Bundle\Bundle;
+
+class FrontendBundle extends Bundle
+{
+}
