+++ WebClient/FrontendBundle/Twig/HelperExtension.php
@@ -1,0 +1,36 @@
+<?php
+
+namespace EMS\ClientHelperBundle\WebClient\Frontend\Twig;
+
+use EMS\ClientHelperBundle\Elasticsearch\ClientRequest;
+use Twig\Extension\AbstractExtension;
+use Twig\TwigFilter;
+
+class HelperExtension extends AbstractExtension
+{
+    /**
+     * {@inheritdoc}
+     */
+    public function getFilters()
+    {
+        return [
+            new TwigFilter('emsch_ouuid', array($this, 'getOuuid')),
+        ];
+    }
+    
+    /**
+     * {@inheritdoc}
+     */
+    public function getOuuid($emsLink)
+    {
+        return ClientRequest::getOuuid($emsLink);
+    }
+
+        /**
+     * {@inheritdoc}
+     */
+    public function getName()
+    {
+        return 'emsch_helper';
+    }
+}
