+++ WebClient/TwigListBundle/Resources/config/config.yml
@@ -1,0 +1,1 @@
+ems_client_helper:    twig_list :        templates:              - {resource: 'EMSClientHelperBundle/WebClient/TwigListBundle', path: 'Resources/views/Pages'}        app_enabled: false        app_base_path : #['Resources/views/default']
