+++ WebClient/TwigListBundle/Resources/config/routing.yml
@@ -1,0 +1,3 @@
+ems_client_helper:
+    resource: '@EMSClientHelperBundle/WebClient/TwigListBundle/Controller/'
+    type : annotation
