+++ WebClient/TwigListBundle/Resources/views/Elements/header.html.twig
@@ -1,0 +1,15 @@
+{# This file is part of the EMSClientHelperBundle #}
+
+<header>
+	<div class="container logo">
+	    <div class="row">
+		      <div class="col-md-7 col-sm-6 col-xs-12" id="logo">
+			        Header 	
+			        <div class="info">
+						<p>'This is the header template. Feel free to override it.' </p>
+				   		<p>To do so, find the original file in ./views/Elements/header.html.twig <br/> and copy it to your project's <span>app/Resources/EMSClientHelperBundle/WebClient/TwigList/views/</span> folder whith respect to the original folder structure</p>
+				   	</div>
+		      </div>
+	    </div>
+    </div>
+</header>
