+++ WebClient/TwigListBundle/Resources/views/base.html.twig
@@ -1,0 +1,20 @@
+{# This file is part of the EMSClientHelperBundle #}
+
+<!DOCTYPE html>
+<html>
+    <head>
+        <meta charset="UTF-8" />
+        <title>{% block title %}Welcome on your base template!{% endblock %}</title>
+        {% block stylesheets %}{% endblock %}
+        <link rel="icon" type="image/x-icon" href="{{ asset('favicon.ico') }}" />
+    </head>
+    <body>
+    	{% block info %} {% endblock %} 
+    	{% include ('TwigListBundle:Elements:header.html.twig') %}
+        
+        {% block body %}{% endblock %}
+        {% block javascripts %}{% endblock %}
+        
+        {% include ('TwigListBundle:Elements:footer.html.twig')  %}
+    </body>
+</html>
