+++ WebClient/TwigListBundle/TwigListBundle.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace EMS\ClientHelperBundle\WeblCient\TwigList;
+
+use Symfony\Component\HttpKernel\Bundle\Bundle;
+
+class TwigListBundle extends Bundle
+{
+}
